requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
cloudinary==1.41.0
celery==5.4.0
redis==5.2.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
greenlet==3.1.1
